+++ src/gui.rs
@@ -17,6 +17,7 @@ mod vm_user_messages_tool;
 mod vm_protobuf_message_list;
 mod vm_server_info_tool;
 mod vm_game_events_tool;
+mod vm_abouthelp;
 // widgets: small tools for displaying common gui components.
 mod w_copyable_field;