+++ src/gui/vm_demo_file.rs
@@ -5,7 +5,7 @@ use super::{
     vm_header_tool::HeaderToolViewModel,
     vm_frames_tool::FramesToolViewModel,
     vm_user_messages_tool::UserMessagesToolViewModel,
-    vm_server_info_tool::ServerInfoViewModel, vm_game_events_tool::GameEventsToolViewModel,
+    vm_server_info_tool::ServerInfoViewModel, vm_game_events_tool::GameEventsToolViewModel, vm_abouthelp::AboutHelpViewModel,
 };
 use source_demo_tool::demo_file::DemoFile;
 use eframe::{
@@ -64,6 +64,11 @@ impl DemoFileViewModel {
 
         let tools: Vec<DemoFileTools> = vec![
             DemoFileTools {
+                name: "?",
+                vm: Box::new(AboutHelpViewModel{}),
+                focus: Focusable::None,
+            },
+            DemoFileTools {
                 name: "Header",
                 vm: Box::new(HeaderToolViewModel::new(header)),
                 focus: Focusable::None,
@@ -93,7 +98,7 @@ impl DemoFileViewModel {
         Self {
             demo_file,
             tools,
-            active_tool_index: 0, // header tool
+            active_tool_index: 1, // header tool
             hover_tool_index: None,
             inner_events: Vec::new(),
             b_inner_events_sent_last: false,