+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ### Latest/Nightly (this branch)
+- Changes
+    - Moved gui initialization to its own thread (stops loading spinner from hitching), closes #24.
+- Internal
+    - set source-demo-tool version to 0.7.2
 
 ### v0.5.2
 - Features